requirements.txt
streamlit
pydub
openai
python-dotenv
whisper
ffmpeg-python
transformers
torch
tqdm
